requirements.txt
asgiref==3.8.1
boto3==1.35.12
botocore==1.35.12
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==7.0.0
Django==5.1.3
django-allauth==65.2.0
django-cors-headers==4.3.0
django-environ==0.11.2
django-filter==23.5
django-rest-swagger==2.2.0
django-storages==1.14.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
mysqlclient==2.1.1
openapi-codec==1.3.2
openpyxl==3.1.5
packaging==23.2
Pillow==10.1.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dateutil==2.9.0.post0
pytz==2023.3.post1
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
s3transfer==0.10.2
simplejson==3.19.3
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
urllib3==2.2.2
whitenoise==6.6.0
xlwt==1.3.0
